// lib/certificates/types.ts
export interface CertificateContent {
    programName: string
    programDates: string
    department?: string
    faculty?: string
    institution?: string
    location?: string
    address?: string
    coordinatorName?: string
    hodName?: string
    footerText?: string
    logoDataUrl?: string
    coordinatorSignatureDataUrl?: string
    hodSignatureDataUrl?: string
}